+++ app/config/global.cpp
@@ -113,4 +113,12 @@ namespace Config {
   bool Global::inhibitSleepWhilePlaying() const {
     return storage.get("inhibit_sleep_while_playing").get<bool>();
   }
+
+  int Global::playlistRowHeight() const {
+    return storage.get("playlist_row_height").get<int>();
+  }
+
+  bool Global::savePlaylistRowHeight(int arg) {
+    return storage.set("playlist_row_height", Config::Value(arg));
+  }
 }