+++ app/config/global.h
@@ -47,6 +47,9 @@ namespace Config {
 
     bool inhibitSleepWhilePlaying() const;
 
+    int playlistRowHeight() const;
+    bool savePlaylistRowHeight(int arg);
+
   private:
     Config::Storage storage;
   };